src/requirements.txt
gym
numpy
matplotlib
pandas
dill
mujoco_py
gitpython
torch
torchvision
GPUtil
tensorboardX
tqdm
sklearn
cox
